refactor: restructure User module
User now has two subpackages: api & impl. Code is currently in impl.
