+++ app/src/main/kotlin/com/potatosheep/kite/app/nav/KiteNavHost.kt
@@ -20,8 +20,8 @@ import com.potatosheep.kite.feature.search.impl.nav.searchScreen
 import com.potatosheep.kite.feature.settings.impl.nav.settingsScreen
 import com.potatosheep.kite.feature.subreddit.impl.nav.navigateToSubreddit
 import com.potatosheep.kite.feature.subreddit.impl.nav.subredditScreen
-import com.potatosheep.kite.feature.user.nav.navigateToUser
-import com.potatosheep.kite.feature.user.nav.userScreen
+import com.potatosheep.kite.feature.user.impl.nav.navigateToUser
+import com.potatosheep.kite.feature.user.impl.nav.userScreen
 import com.potatosheep.kite.feature.video.nav.navigateToVideo
 import com.potatosheep.kite.feature.video.nav.videoScreen
