+++ feature/user/impl/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    namespace = "com.potatosheep.kite.feature.user"
+    namespace = "com.potatosheep.kite.feature.user.impl"
 }
 
 dependencies {
