+++ feature/user/impl/src/main/kotlin/com/potatosheep/kite/feature/user/impl/nav/UserNavigation.kt
@@ -1,24 +1,18 @@
-package com.potatosheep.kite.feature.user.nav
+package com.potatosheep.kite.feature.user.impl.nav
 
 import androidx.compose.animation.AnimatedContentTransitionScope
-import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.scaleIn
 import androidx.compose.animation.scaleOut
-import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
-import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import com.potatosheep.kite.core.common.enums.SortOption
-import com.potatosheep.kite.feature.user.UserRoute
+import com.potatosheep.kite.feature.user.impl.UserRoute
 import kotlinx.serialization.Serializable
-import kotlin.reflect.typeOf
 
 @Serializable
 data class UserNav(val user: String)
