+++ feature/user/impl/src/main/res/mipmap-anydpi/ic_launcher.xml
