+++ settings.gradle.kts
@@ -43,7 +43,8 @@ include(":feature:subreddit:impl")
 include(":feature:subreddit:api")
 include(":core:common")
 include(":feature:subredditwiki")
-include(":feature:user")
+include(":feature:user:impl")
+include(":feature:user:api")
 include(":feature:search:impl")
 include(":feature:search:api")
 include(":core:markdown")
